refactor(cli): split orchestration handlers
